update: move constants and interfaces to dedicated files, signAccountOp
