+++ src/interfaces/signAccountOp.ts
@@ -1,0 +1,9 @@
+type Warning = {
+  id: string
+  title: string
+  text: string
+  promptBeforeSign: boolean
+  displayBeforeSign: boolean
+}
+
+export type { Warning }
